fix(data-vi): fix issues when using external oracle data source for data visualization
