+++ packages/plugins/@nocobase/plugin-data-visualization/src/server/__tests__/query.test.ts
@@ -13,13 +13,14 @@ import { vi } from 'vitest';
 import {
   cacheMiddleware,
   checkPermission,
-  parseBuilder,
   parseFieldAndAssociations,
   parseVariables,
   postProcess,
 } from '../actions/query';
 import { Database } from '@nocobase/database';
 import * as formatter from '../formatter';
+import { createQueryParser } from '../query-parser';
+import { QueryParser } from '../query-parser/query-parser';
 
 describe('query', () => {
   describe('parseBuilder', () => {
@@ -157,7 +158,8 @@ describe('query', () => {
           },
         },
       };
-      await compose([parseFieldAndAssociations, parseBuilder])(context, async () => {});
+      const queryParser = createQueryParser(db);
+      await compose([parseFieldAndAssociations, queryParser.parse()])(context, async () => {});
       expect(context.action.params.values.queryParams.attributes).toEqual([
         [db.sequelize.col('orders.price'), 'price'],
       ]);
@@ -179,7 +181,7 @@ describe('query', () => {
           },
         },
       };
-      await compose([parseFieldAndAssociations, parseBuilder])(context2, async () => {});
+      await compose([parseFieldAndAssociations, queryParser.parse()])(context2, async () => {});
       expect(context2.action.params.values.queryParams.attributes).toEqual([
         [db.sequelize.fn('sum', db.sequelize.col('orders.price')), 'price-alias'],
       ]);
@@ -203,20 +205,17 @@ describe('query', () => {
           },
         },
       };
+      const queryParser = createQueryParser(db);
       try {
-        await compose([parseFieldAndAssociations, parseBuilder])(context, async () => {});
+        await compose([parseFieldAndAssociations, queryParser.parse()])(context, async () => {});
       } catch (error) {
         expect(error.message).toBe('Invalid aggregation function: if(1=2,sleep(1),sleep(3)) and sum');
       }
     });
 
     it('should parse dimensions', async () => {
-      vi.spyOn(formatter, 'createFormatter').mockImplementation(
-        () =>
-          ({
-            format: () => 'formatted-field' as any,
-          }) as any,
-      );
+      const queryParser = createQueryParser(db);
+      vi.spyOn(queryParser.formatter, 'format').mockImplementation(() => 'formatted-field' as any);
       const dimensions = [
         {
           field: ['createdAt'],
@@ -235,7 +234,7 @@ describe('query', () => {
           },
         },
       };
-      await compose([parseFieldAndAssociations, parseBuilder])(context, async () => {});
+      await compose([parseFieldAndAssociations, queryParser.parse()])(context, async () => {});
       expect(context.action.params.values.queryParams.attributes).toEqual([['formatted-field', 'Created at']]);
       expect(context.action.params.values.queryParams.group).toEqual([]);
       const measures = [
@@ -256,7 +255,7 @@ describe('query', () => {
           },
         },
       };
-      await compose([parseFieldAndAssociations, parseBuilder])(context2, async () => {});
+      await compose([parseFieldAndAssociations, queryParser.parse()])(context2, async () => {});
       expect(context2.action.params.values.queryParams.group).toEqual(['formatted-field']);
     });
 
@@ -277,7 +276,8 @@ describe('query', () => {
           },
         },
       };
-      await compose([parseFieldAndAssociations, parseBuilder])(context, async () => {});
+      const queryParser = createQueryParser(db);
+      await compose([parseFieldAndAssociations, queryParser.parse()])(context, async () => {});
       expect(context.action.params.values.queryParams.where.createdAt).toBeDefined();
     });
 
@@ -332,7 +332,6 @@ describe('query', () => {
       await parseVariables(context, async () => {});
       const { filter } = context.action.params.values;
       const dateOn = filter.$and[0].createdAt.$dateOn;
-      console.log(dateOn);
       expect(new Date(dateOn).getTime()).toBeLessThanOrEqual(new Date().getTime());
       const userId = filter.$and[1].userId.$eq;
       expect(userId).toBe(1);
